Revert True Append to named pipe
In an attempted fix, @stevenweaver switched from a named pipe to a temporary file. However, I think the bug was elsewhere (I had forgotten to close the output file stream; I don't think the issue actually ended up having anything to do with the named pipe). I have reverted some of the changes to go back to the prior logic that used the named pipe.